Add support for the output.
The output objects are now observing the collector and are notified as soon as
something happens there, using their "push" method.

I'm not sure that's actually the best way to send them information, feel free
to change this if you can find something better, but that will do for now.