Use cycle and current_cycle properly.
When you specify cycles, you provide values such as 10:20:30 which means that
you want to have 3 cycles, with 3 different number of tests to run.

In loads, "cycle" is the cycle you are in (10, 20 or 30), whereas current_cycle
is the identifier of the test you're running, that's somehow confusing and
I was actually confused.

Docs needs an update as well.